+++ CHANGELOG.md
@@ -12,6 +12,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Variants query backend allows rank_score filtering
 - Added script to tabulate causatives clinical filter rank
 - Do not display inheritance models associated to ORPHA terms on variant page
+- From gene panel page, when exporting a panel to .txt or .pdf, export all fields instead of the first columns only
 
 ## [4.89.2]
 ## Fixed